requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
crispy-bootstrap5==0.6
Django==3.2.9
django-bootstrap-v5==1.0.7
django-crispy-forms==1.13.0
django-phonenumber-field==6.0.0
django-softdelete==0.9.2
fontawesome-free==5.15.4
mysqlclient==2.1.0
numpy==1.21.5
pdfkit==0.6.0
phonenumberslite==8.12.37
Pillow==8.4.0
pytz==2022.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
